+++ CHANGELOG.md
@@ -22,8 +22,10 @@ Given the `MAJOR.MINOR.PATCH` pattern, here is how we decide to increment:
 ### Changed
 
 - JobListingPages are now only associated with a single region.
+- Updated custom Wagtail admin templates to 1.7 version.
 
 ### Removed
+- Can no longer delete via `/delete` in Wagtail
 
 
 ### Fixed